tweaks per NickSh code review - consolidate code where it's actually being used and has access to Skyline resx file for best error handling
